Improve AttritubedAction and Tasks
